+++ R/generics.R
@@ -285,14 +285,15 @@ setGeneric("reportCSV", function(fGroups, path = "report", reportFGroupsAsRows =
                                  compsCluster = NULL, components = NULL,
                                  retMin = TRUE, clearPath = FALSE) standardGeneric("reportCSV"))
 setGeneric("reportPDF", function(fGroups, path = "report", reportFGroups = TRUE,
-                                 formulas = NULL, reportFormulaSpectra = TRUE, compounds = NULL, compoundNormalizeScores = "max",
+                                 formulas = NULL, formulasTopMost = 5, reportFormulaSpectra = TRUE,
+                                 compounds = NULL, compoundNormalizeScores = "max",
                                  compoundExclNormScores = c("score", "individualMoNAScore"),
                                  compoundOnlyUsedScorings = TRUE, compoundTopMost = 5,
                                  compsCluster = NULL, components = NULL, MSPeakLists = NULL, retMin = TRUE,
                                  EICGrid = c(2, 1), EICRtWindow = 20, EICMzWindow = 0.005, EICTopMost = NULL,
                                  EICOnlyPresent = TRUE, clearPath = FALSE) standardGeneric("reportPDF"))
 setGeneric("reportMD", function(fGroups, path = "report", reportPlots = c("chord", "venn", "upset", "eics", "formulas"),
-                                formulas = NULL, compounds = NULL, compoundNormalizeScores = "max",
+                                formulas = NULL, formulasTopMost = 5, compounds = NULL, compoundNormalizeScores = "max",
                                 compoundExclNormScores = c("score", "individualMoNAScore"),
                                 compoundOnlyUsedScorings = TRUE, compoundTopMost = 5, compsCluster = NULL,
                                 includeMFWebLinks = "compounds", components = NULL, interactiveHeat = FALSE,