+++ R/main.R
@@ -84,8 +84,9 @@ NULL
 #'   cores available. Default is set by \code{\link[parallel]{detectCores}}.
 #'
 #'   \item \code{patRoon.path.pwiz}: The path in which the
-#'   \command{ProteoWizard} binaries are installed. Only need to be set if not
-#'   yet in \option{PATH}.
+#'   \command{ProteoWizard} binaries are installed. If unset an attempt is made
+#'   to find this directory from the Windows registry and \option{PATH}
+#'   environment variable.
 #'
 #'   \item \code{patRoon.path.GenForm}: The path to the \command{GenForm}
 #'   executable. If not set (the default) the internal \code{GenForm} binary is